[REFACTOR] Move User Package
