+++ src/main/java/com/hackathon/MYD/model/user/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.hackathon.MYD.model.repository;
+package com.hackathon.MYD.model.user.repository;
 
 import com.hackathon.MYD.model.user.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
